ref: move virtual_text out of helpers
